# fix: remove ScaleF correction to honor configured cutoff frequency