AI Generated code. Claude Sonnet 4.5 (Preview) via VSCode Copilot.
<hr>
Purpose: To accurately set cutoff as per user-configuration.
<hr>

### **For version 257, Pilot would change to these effective cutoffs if wishing to match version 256:**
 - example, i was PT2@90hz on IMUF 256 -- if i want the exact same, then i should set IMUF 257 PT@140hz. However, I  always thought 90hz was too low, and analysis proved it was not really 90hz, but instead 140hz.  Now i can look at blackbox/graphs and discern the **_proper_** cutoff to set.

### IMUF 256 (Old Code with ScaleF) - Effective Cutoff Frequencies

| Configured | PT1 Effective | PT2 Effective | PT3 Effective |
|------------|---------------|---------------|---------------|
| 70 Hz      | 70 Hz         | 109 Hz        | 137 Hz        |
| 90 Hz      | 90 Hz         | 140 Hz        | 176 Hz        |
| 110 Hz     | 110 Hz        | 171 Hz        | 216 Hz        |
| 130 Hz     | 130 Hz        | 202 Hz        | 255 Hz        |
| 150 Hz     | 150 Hz        | 233 Hz        | 294 Hz        |
| 170 Hz     | 170 Hz        | 264 Hz        | 333 Hz        |
| 190 Hz     | 190 Hz        | 295 Hz        | 373 Hz        |
| 210 Hz     | 210 Hz        | 326 Hz        | 412 Hz        |
| 230 Hz     | 230 Hz        | 357 Hz        | 451 Hz        |
| 250 Hz     | 250 Hz        | 389 Hz        | 491 Hz        |
| 270 Hz     | 270 Hz        | 420 Hz        | 530 Hz        |
| 290 Hz     | 290 Hz        | 451 Hz        | 569 Hz        |

**ScaleF multipliers:**
- PT1: 1.0× (no scaling)
- PT2: 1.554×
- PT3: 1.961×

<hr>

## Architecture Explanation and Reasoning for this PR:

### Hardware Sample Rate vs Output Rate

The IMU-F has two different "rates" that are often confused:

1. **Hardware Gyro Sample Rate** (fixed at 32 kHz)
   - Set by `gyroConfig.rateDiv = 1` in `gyro_device.c`
   - Gyro hardware samples at 32,000 Hz
   - Never changes at runtime

2. **Output Rate to Flight Controller** (user-configurable: 32k, 16k, 8k, 4k, etc.)
   - Controlled by `gyroSettingsConfig.rate` → `loopDivider`
   - Only affects how often data is SENT to the FC
   - Does NOT affect filtering rate

### Data Flow

```
┌─────────────────────────────────────────────────────────────┐
│ Every 31.25 μs (32 kHz) - triggered by gyro hardware:      │
├─────────────────────────────────────────────────────────────┤
│ 1. Gyro interrupt fires (gyroDataReadDone = 1)             │
│ 2. scheduler_run() executes:                               │
│    ├─ gyro_int_to_float()   ← Convert raw gyro data        │
│    ├─ run_gyro_filters()    ← **FILTER ALWAYS RUNS HERE**  │
│    │   └─ filter_data()                                    │
│    │      ├─ kalman_update()                               │
│    │      └─ ptnFilterApply() ← Uses REFRESH_RATE=32kHz   │
│    ├─ update_quaternions()                                 │
│    └─ fire_spi_send_ready() ← Uses loopDivider            │
│         └─ Only sends data every Nth call                  │
└─────────────────────────────────────────────────────────────┘
```

### Code Evidence

**`scheduler.c:10-15`:**
```c
inline void scheduler_run(void)
{
    while(!gyroDataReadDone); // Wait for 32kHz gyro interrupt
    gyro_int_to_float(&gyroRxFrame);
    run_gyro_filters();  // ← ALWAYS CALLED AT 32 KHZ
    // ...
    fire_spi_send_ready(); // ← Uses loopDivider to skip sends
}
```

**`gyro.c:309-312`:**
```c
void run_gyro_filters(void)
{
    filter_data(&rawRateData, &rawAccData, gyroTempData, &filteredData);
}
```

**`filter.c:88-91`:**
```c
filteredData->rateData.x = ptnFilterApply(filteredData->rateData.x, &(lpfFilterStateRate.x));
filteredData->rateData.y = ptnFilterApply(filteredData->rateData.y, &(lpfFilterStateRate.y));
filteredData->rateData.z = ptnFilterApply(filteredData->rateData.z, &(lpfFilterStateRate.z));
```

**`ptnFilter.c:15`:**
```c
filter->k = REFRESH_RATE / ((1.0f / (2.0f * M_PI_FLOAT * f_cut)) + REFRESH_RATE);
// REFRESH_RATE = 0.00003125f = 1/32000 Hz
```

### What loopDivider Actually Controls

**`gyro.c:418-420`:**
```c
if (everyOther-- <= 0)
{
    everyOther = loopDivider; // Reset counter
    // ... SEND data to flight controller
}
```

The `loopDivider` is a **skip counter** for OUTPUT:
- `loopDivider = 0` → send every time (32 kHz output)
- `loopDivider = 1` → send every 2nd time (16 kHz output)  
- `loopDivider = 3` → send every 4th time (8 kHz output)
- `loopDivider = 7` → send every 8th time (4 kHz output)

**The filter runs BEFORE this check**, so it always operates at 32 kHz.

---

## Why This Is Correct

### 1. Anti-Aliasing Benefits

By filtering at the hardware sample rate (32 kHz), the filter acts as an anti-aliasing filter regardless of output rate:

- **8 kHz output:** Filter prevents aliasing of frequencies above 4 kHz
- **4 kHz output:** Filter prevents aliasing of frequencies above 2 kHz
- **2 kHz output:** Filter prevents aliasing of frequencies above 1 kHz

This is the CORRECT approach! If the filter ran at the output rate, you'd get aliasing artifacts.

### 2. Constant Filter Coefficients

The filter coefficient `k` depends only on:
- Configured cutoff frequency (`f_cut`)
- Hardware sample rate (`REFRESH_RATE = 32 kHz`)

Since both are constant, the filter behavior is consistent regardless of output rate.

### 3. No Nyquist Violations

With a 32 kHz sample rate:
- Nyquist frequency: 16 kHz
- Typical cutoff range: 50-200 Hz
- Filter is operating well below Nyquist limit ✅

---

## Verification

### Test Case: 90 Hz Cutoff, PT2 Filter

**Scenario 1: 32 kHz output rate**
- Filter runs at: 32 kHz ✓
- k coefficient calculated with: REFRESH_RATE = 0.00003125f ✓
- Per-stage cutoff: 90 Hz ✓
- Combined -3dB point: ~58 Hz ✓

**Scenario 2: 4 kHz output rate**
- Filter runs at: 32 kHz ✓ (still!)
- k coefficient: Same as above ✓
- Per-stage cutoff: 90 Hz ✓
- Combined -3dB point: ~58 Hz ✓
- Output decimated 8:1 after filtering ✓

**Result:** Filter behavior is IDENTICAL in both cases.

---

## Potential Issues (None Found)

### Could the filter be running at output rate? ❌

**No.** The filter is called from `scheduler_run()` which executes every gyro interrupt (32 kHz). The `loopDivider` is only used in `fire_spi_send_ready()` which comes AFTER filtering.

### Could REFRESH_RATE be wrong? ❌

**No.** The gyro hardware is configured with:
- `gyroConfig.rateDiv = 1`
- `SMPLRT_DIV = rateDiv - 1 = 0`
- Internal sample rate = 32 kHz / (1 + 0) = 32 kHz ✓

### Could different gyro rates need different filter coefficients? ❌

**No.** Since the filter always runs at 32 kHz, the coefficients are always correct.

---

## Conclusion

✅ **The implementation is correct and robust.**

The hardcoded `REFRESH_RATE = 0.00003125f` (32 kHz) is appropriate because:
1. The gyro hardware always samples at 32 kHz
2. The filter always runs at 32 kHz
3. The output rate only affects data transmission, not filtering

**No changes needed.** The fix works correctly regardless of user-configured output rate.

---

## Additional Notes

### If Hardware Sample Rate Ever Changes

If future hardware or firmware changes the actual gyro sample rate, then `REFRESH_RATE` would need to be updated or made dynamic. But currently:
- `gyroConfig` is `const` (compile-time constant)
- `rateDiv = 1` is hardcoded
- No runtime modification of sample rate exists

### Comments in Code

The commented-out `HALF_GYRO_DT` values in `imu.c` show that the 32 kHz rate has been stable:
```c
//#define HALF_GYRO_DT 0.000125f  //1/2 of dT for gyro sample rate which is 32 KHz
//#define HALF_GYRO_DT 0.00025f   //1/2 of dT for gyro sample rate which is 32 KHz
#define HALF_GYRO_DT 0.000015625f //1/2 of dT for gyro sample rate which is 32 KHz
```

The final value (0.000015625f = 1/(2×32000)) confirms 32 kHz operation.

---

**Date:** October 4, 2025  
**Analysis by:** GitHub Copilot  
**Issue Reference:** emuflight/imu-f#15